src/requirements.txt
amqp==5.2.0
anyio==3.7.1
api-client==1.3.1
asgiref==3.7.2
astroid==3.1.0
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
billiard==4.2.0
celery==5.3.6
celery-stubs==0.1.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.3
cron-descriptor==1.4.3
cssbeautifier==1.15.1
dill==0.3.8
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.10
django-celery-beat==2.5.0
django-dynamic-fixture @ git+https://github.com/paulocheque/django-dynamic-fixture.git@eb4424798bb6ba60c3eab5ca1a23f86891051516
django-object-actions==4.2.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-stubs==3.14.5
djlint==1.34.1
drf-spectacular==0.27.1
drf-test-generator==0.0.5
drf-yasg==1.21.7
EditorConfig==0.12.4
filelock==3.13.1
flower==2.0.1
ghp-import==2.1.0
graphql-core==3.2.3
griffe==0.41.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httpcore==1.0.4
httpx==0.27.0
humanize==4.9.0
hypothesis==6.98.15
hypothesis-graphql==0.11.0
hypothesis-jsonschema==0.23.1
identify==2.5.35
idna==3.6
inflection==0.3.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.3
jsbeautifier==1.15.1
json5==0.9.17
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
junit-xml==1.9
kombu==5.3.5
lazy-object-proxy==1.10.0
Markdown==3.5.2
MarkupSafe==2.1.5
mccabe==0.7.0
mergedeep==1.3.4
mkautodoc==0.2.0
mkdocs==1.5.3
mkdocs-autorefs==1.0.0
mkdocs-gen-files==0.5.0
mkdocs-literate-nav==0.6.1
mkdocs-material==9.5.12
mkdocs-material-extensions==1.3.1
mkdocs-section-index==0.3.8
mkdocstrings==0.24.1
mkdocstrings-python==1.8.0
multidict==6.0.5
mypy==1.7.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
paginate==0.5.6
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
Pygments==2.17.2
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pymdown-extensions==10.7
pyrate-limiter==2.10.0
pytest==7.4.4
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-drf==1.1.3
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
pytest-mock==3.12.0
pytest-subtests==0.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
redis==5.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
schemathesis==3.25.5
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
starlette==0.37.1
starlette-testclient==0.3.0
swagger-spec-validator==3.0.3
tenacity==8.2.3
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.12.4
tornado==6.4
tqdm==4.66.2
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240218
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.1
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
